+++ tests/testthat/test-col-spec.R
@@ -82,9 +82,10 @@ test_that("print(col_spec) with collector_skip", {
 
 test_that("print(col_spec) with truncated output", {
   out <- col_spec_standardise("a,b,c\n1,2,3", col_types = cols(.default = "c"))
-  expect_output(print(out, n = 2),
+  expect_output(print(out, n = 2, condense = FALSE),
     regex_escape(
-"cols(.default = col_character(),
+"cols(
+  .default = col_character(),
   a = col_character(),
   b = col_character()
   # ... with 1 more columns
@@ -145,4 +146,21 @@ test_that("print(cols_only()) prints properly", {
 test_that("print(col_spec) with n == 0 prints nothing", {
   out <- col_spec_standardise("a,b,c\n1,2,3")
   expect_silent(print(out, n = 0))
+})
+
+test_that("print(col_spec, condense = TRUE) condenses the spec", {
+  out <- col_spec_standardise("a,b,c,d\n1,2,3,a")
+  expect_output(print(cols_condense(out)),
+    regex_escape(
+"cols(
+  .default = col_integer(),
+  d = col_character()
+)"))
+
+  out <- col_spec_standardise("a,b,c,d\n1,2,3,4")
+  expect_output(print(cols_condense(out)),
+    regex_escape(
+"cols(
+  .default = col_integer()
+)"))
 })