+++ NEWS.md
@@ -3,6 +3,9 @@
 * `parse_number()` is slightly more flexible - it now parses numbers up
   to the first ill-formed character. For example `parse_number("-3-")`
   and `parse_number("...3...")` now return -3 and 3 respectively.
+* `condense_spec()` function to condense a column specification, useful when
+* `cols_condense()` function to condense a column specification, useful when
+  only a handful of columns are of a different type (#466, @jimhester).
 
 * New `stop_for_problems(x)` throws an error if `x` had any parsing problems 
   (#465).